# Add and use wibox.hierarchy:update()