@actionless 2015-09-21
added one more test to benchmark,

before this patch:
```
        create wibox: 0.0677603  sec/iter ( 15 iters, 1.083 sec for benchmark)
    update textclock: 0.0036652  sec/iter (300 iters, 2.211 sec for benchmark)
  relayout textclock: 0.00253013 sec/iter (418 iters, 1.06 sec for benchmark)
    redraw textclock: 0.00114886 sec/iter (895 iters, 1.881 sec for benchmark)
          tag switch: 0.00419569 sec/iter (252 iters, 3.561 sec for benchmark)
```

after:
```
        create wibox: 0.0675195  sec/iter ( 16 iters, 1.146 sec for benchmark)
    update textclock: 0.00172781 sec/iter (693 iters, 2.347 sec for benchmark)
  relayout textclock: 0.000477135 sec/iter (2153 iters, 2.016 sec for benchmark)
    redraw textclock: 0.00107928 sec/iter (953 iters, 2.796 sec for benchmark)
          tag switch: 0.00442825 sec/iter (228 iters, 1.628 sec for benchmark)
```