@psychon 2015-09-22
Wow. We can update the textclock faster than switch tags? That's unexpected. However, as mentioned in #465, I don't think your new benchmark measures what you think it measures (and if it would, I'd expect worse numbers, since the taglist is currently updated in a quite inefficient way (the code removes everything from a fixed layout and re-adds things again, thus forcing a full relayout of all the widgets in there. At least I think that we can skip unneeded redraws since in the end nothing actually changed.)