@actionless 2015-09-22
update the test:

before:
```
        create wibox: 0.0668394  sec/iter ( 16 iters, 1.136 sec for benchmark)
    update textclock: 0.003537   sec/iter (328 iters, 2.227 sec for benchmark)
  relayout textclock: 0.00244165 sec/iter (441 iters, 1.079 sec for benchmark)
    redraw textclock: 0.00115537 sec/iter (920 iters, 1.901 sec for benchmark)
          tag switch: 0.208276   sec/iter (  5 iters, 1.287 sec for benchmark)
```

after:
```
        create wibox: 0.0676614  sec/iter ( 16 iters, 1.149 sec for benchmark)
    update textclock: 0.00157576 sec/iter (712 iters, 2.258 sec for benchmark)
  relayout textclock: 0.000480156 sec/iter (2208 iters, 1.061 sec for benchmark)
    redraw textclock: 0.000930917 sec/iter (1143 iters, 1.842 sec for benchmark)
          tag switch: 0.238446   sec/iter (  5 iters, 1.425 sec for benchmark)
```

still some "floatnig" difference, about 0.02..0.06