@blueyed 2015-09-23
Worked fine for me, will merge it.